requirements.txt
streamlit
pandas
plotly
folium
streamlit-folium
Pillow